Fix integer truncation in StreamReader size calculations

+++ include/assimp/StreamReader.h
@@ -182,7 +182,7 @@ public:
     // ---------------------------------------------------------------------
     /// Get the remaining stream size (to the end of the stream)
     size_t GetRemainingSize() const {
-        return (unsigned int)(mEnd - mCurrent);
+        return static_cast<size_t>(mEnd - mCurrent);
     }
 
     // ---------------------------------------------------------------------
@@ -190,7 +190,7 @@ public:
      *  return value is the remaining size of the stream if no custom
      *  read limit has been set. */
     size_t GetRemainingSizeToLimit() const {
-        return (unsigned int)(mLimit - mCurrent);
+        return static_cast<size_t>(mLimit - mCurrent);
     }
 
     // ---------------------------------------------------------------------
@@ -234,7 +234,7 @@ public:
 
     /// @brief Get the current offset from the beginning of the file
     int GetCurrentPos() const {
-        return (unsigned int)(mCurrent - mBuffer);
+        return static_cast<int>(mCurrent - mBuffer);
     }
 
     void SetCurrentPos(size_t pos) {
@@ -244,7 +244,7 @@ public:
     // ---------------------------------------------------------------------
     /** Setup a temporary read limit
      *
-     *  @param limit Maximum number of bytes to be read from
+     *  @param _limit Maximum number of bytes to be read from
      *    the beginning of the file. Specifying UINT_MAX
      *    resets the limit to the original end of the stream.
      *  Returns the previously set limit. */
